Extract design post parser and split into sub-components
Breaks the God component at design/[slug]/+page.svelte into:
- lib/utils/parse-design-markdown.ts: parsing logic + block splitting
- lib/components/shared/design-post-header.svelte: title + meta
- lib/components/shared/design-post-gallery.svelte: image gallery

The page component is now a thin orchestrator (~30 lines vs ~155).
Parsing is independently testable, and gallery/header can be
extended or replaced without touching each other.
